+++ metadata-ingestion/src/datahub/ingestion/source/ge_data_profiler.py
@@ -5,14 +5,11 @@ import itertools
 import logging
 import os
 import threading
-import traceback
 import unittest.mock
 import uuid
-from typing import Any, ClassVar, Dict, Iterable, Iterator, List, Optional, Tuple, Union
+from typing import Any, Dict, Iterable, Iterator, List, Optional, Tuple, Union
 
-import greenlet
 import pydantic
-import sqlalchemy.engine
 from great_expectations.data_context import BaseDataContext
 from great_expectations.data_context.types.base import (
     DataContextConfig,
@@ -37,6 +34,7 @@ from datahub.metadata.schema_classes import (
     ValueFrequencyClass,
 )
 from datahub.utilities.perf_timer import PerfTimer
+from datahub.utilities.sqlalchemy_query_combiner import SQLAlchemyQueryCombiner
 
 logger: logging.Logger = logging.getLogger(__name__)
 
@@ -155,146 +153,6 @@ class GEProfilingConfig(ConfigModel):
                 values[max_num_fields_to_profile_key] = 10
 
         return values
-
-
-@dataclasses.dataclass
-class _QueryFuture:
-    conn: Connection
-    query: Any
-    multiparams: Any
-    params: Any
-
-    done: bool = False
-    res: Any = None
-
-
-@dataclasses.dataclass
-class _SQLAlchemyQueryCoalescenceManager:
-    # TODO this must be a singleton, should not be set up per-thread
-
-    # Without the staticmethod decorator, Python thinks that this is an instance
-    # method and attempts to bind self to it.
-    _underlying_sa_execute_method: ClassVar = staticmethod(
-        sqlalchemy.engine.Connection.execute
-    )
-
-    _allowed_single_row_query_methods: ClassVar = [
-        (
-            "great_expectations/dataset/sqlalchemy_dataset.py",
-            {
-                "get_row_count",
-                "get_column_min",
-                "get_column_max",
-                "get_column_mean",
-                "get_column_median",
-                "get_column_stdev",
-                "get_column_stdev",
-                "get_column_nonnull_count",
-                "get_column_unique_count",
-            },
-        ),
-    ]
-
-    catch_exceptions: bool
-
-    # There will be one main greenlet per thread. As such, queries will be
-    # queued according to the main greenlet's thread ID.
-    _queries_by_thread: Dict[greenlet.greenlet, List[_QueryFuture]] = {}
-
-    def _is_single_row_query_method(
-        self, stack: traceback.StackSummary, query: Any
-    ) -> bool:
-        # We'll do this the inefficient way since the arrays are pretty small.
-        for frame in stack:
-            for file_suffix, allowed_methods in self._allowed_single_row_query_methods:
-                if not frame.filename.endswith(file_suffix):
-                    continue
-                if frame.name in allowed_methods:
-                    return True
-        return False
-
-    def _get_main_greenlet(self) -> greenlet.greenlet:
-        let = greenlet.getcurrent()
-        while let.parent is not None:
-            let = let.parent
-        return let
-
-    def _get_queue(self, main_greenlet: greenlet.greenlet) -> List[_QueryFuture]:
-        assert main_greenlet.parent is None
-
-        # Because of the GIL, this operation is thread-safe. Hence, we can
-        # just add the main greenlet here without any special consideration.
-        # https://stackoverflow.com/a/6953515/5004662
-        # https://docs.python.org/3/glossary.html#term-global-interpreter-lock
-
-        return self._queries_by_thread.setdefault(main_greenlet, [])
-
-    def _handle_execute(
-        self, conn: Connection, query: Any, multiparams: Any, params: Any
-    ) -> Tuple[bool, Any]:
-        # Returns True with result if the query was handled, False if it
-        # should be executed normally using the fallback method.
-
-        # Must handle synchronously if the query was issued from the main greenlet.
-        main_greenlet = self._get_main_greenlet()
-        if greenlet.getcurrent() == main_greenlet:
-            breakpoint()
-            return False, None
-
-        # Don't attempt to handle if these are set.
-        if multiparams or params:
-            return False, None
-
-        # Attempt to match against the known single-row query methods.
-        stack = traceback.extract_stack()
-        if not self._is_single_row_query_method(stack, query):
-            return False, None
-
-        # Figure out how many columns this query returns.
-        # TODO add escape hatch
-        if not hasattr(query, "columns"):
-            return False, None
-        columns = list(query.columns)
-        assert len(columns) > 0
-
-        # Add query to the queue.
-        queue = self._get_queue(main_greenlet)
-        query_future = _QueryFuture(conn, query, multiparams, params)
-        queue.append(query_future)
-
-        # Yield control back to the main greenlet until the query is done.
-        # We assume that the main greenlet will be the one that actually executes the query.
-        while not query_future.done:
-            main_greenlet.switch()
-
-        return True, query_future.res
-
-    @contextlib.contextmanager
-    def activate(self) -> Iterator["_SQLAlchemyQueryCoalescenceManager"]:
-        def _sa_execute_fake(conn, query, *args, **kwargs):
-            try:
-                handled, result = self._handle_execute(conn, query, args, kwargs)
-            except Exception as e:
-                if not self.catch_exceptions:
-                    raise e
-                logger.exception(
-                    f"Failed to execute query normally, using fallback: {str(query)}"
-                )
-                return self._underlying_sa_execute_method(conn, query, *args, **kwargs)
-            else:
-                if handled:
-                    logger.info(f"Query was handled: {str(query)}")
-                    return result
-                else:
-                    logger.info(f"Executing query normally: {str(query)}")
-                    return self._underlying_sa_execute_method(
-                        conn, query, *args, **kwargs
-                    )
-
-        with unittest.mock.patch(
-            "sqlalchemy.engine.Connection.execute", _sa_execute_fake
-        ):
-            yield self
 
 
 @dataclasses.dataclass
@@ -652,7 +510,7 @@ class DatahubGEProfiler:
     def generate_profiles(
         self, requests: List[GEProfilerRequest], max_workers: int
     ) -> Iterable[Tuple[GEProfilerRequest, Optional[DatasetProfileClass]]]:
-        with PerfTimer() as timer, _SQLAlchemyQueryCoalescenceManager(
+        with PerfTimer() as timer, SQLAlchemyQueryCombiner(
             catch_exceptions=self.config.catch_exceptions
         ).activate() as query_combiner:
             max_workers = min(max_workers, len(requests))
@@ -683,7 +541,7 @@ class DatahubGEProfiler:
 
     def _generate_profile_from_request(
         self,
-        query_combiner: _SQLAlchemyQueryCoalescenceManager,
+        query_combiner: SQLAlchemyQueryCombiner,
         request: GEProfilerRequest,
     ) -> Tuple[GEProfilerRequest, Optional[DatasetProfileClass]]:
         return request, self._generate_single_profile(
@@ -694,7 +552,7 @@ class DatahubGEProfiler:
 
     def _generate_single_profile(
         self,
-        query_combiner: _SQLAlchemyQueryCoalescenceManager,
+        query_combiner: SQLAlchemyQueryCombiner,
         pretty_name: str,
         schema: str = None,
         table: str = None,
